+++ android/app/src/main/java/com/woowacourse/ody/presentation/common/listener/NextListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.common.listener
+
+interface NextListener {
+    fun onNext()
+}
