+++ android/app/src/main/java/com/woowacourse/ody/presentation/intro/IntroViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.woowacourse.ody.presentation.common.Event
 import com.woowacourse.ody.presentation.common.emit
+import com.woowacourse.ody.presentation.intro.listener.IntroListener
 
 class IntroViewModel : ViewModel(), IntroListener {
     private val _navigateAction: MutableLiveData<Event<IntroNavigateAction>> = MutableLiveData(null)
