+++ android/app/src/main/java/com/woowacourse/ody/presentation/intro/listener/IntroListener.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.intro
+package com.woowacourse.ody.presentation.intro.listener
 
 interface IntroListener {
     fun onClickInputInviteCodeButton()
