+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/listener/ShareListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.meetingroom.listener
+
+interface ShareListener {
+    fun share()
+}
