+++ android/app/src/main/res/layout/activity_join_info.xml
@@ -11,11 +11,11 @@
 
         <variable
             name="nextListener"
-            type="com.woowacourse.ody.presentation.common.NextListener" />
+            type="com.woowacourse.ody.presentation.common.listener.NextListener" />
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.common.BackListener" />
+            type="com.woowacourse.ody.presentation.common.listener.BackListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
