+++ android/app/src/main/res/layout/activity_meeting_info.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.common.BackListener" />
+            type="com.woowacourse.ody.presentation.common.listener.BackListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
