+++ android/app/src/main/res/layout/toolbar_back.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="backListener"
-            type="com.woowacourse.ody.presentation.common.BackListener" />
+            type="com.woowacourse.ody.presentation.common.listener.BackListener" />
     </data>
 
     <androidx.appcompat.widget.Toolbar
